% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_individual_input_options}
\alias{get_individual_input_options}
\title{Get input options for individual job UI}
\usage{
get_individual_input_options()
}
\value{

}
\description{
Get input options for individual job UI
}
